+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+- API to ASCII encode (armor) KeyRings:
+	```go
+	func (keyRing *KeyRing) Armor() (string, error)
+	```
+
 ## [2.8.0-alpha.1] 2024-04-09
 
 ### Added